updating faq tests
