+++ package.json
@@ -31,7 +31,7 @@
     "react-app",
     "bin/cli",
     "mocks",
-    "test/e2e"
+    "test"
   ],
   "license": "CC0-1.0",
   "homepage": "https://github.com/Enterprise-CMCS/macpro-mako#readme",
