+++ playwright/.auth/state-user.json
@@ -1,0 +1,80 @@
+{
+  "cookies": [
+    {
+      "name": "csrf-state",
+      "value": "\"\"",
+      "domain": "playwright-launchdarkly-login-2a6h5ekup131c0ajka6molsjju.auth.us-east-1.amazoncognito.com",
+      "path": "/",
+      "expires": 1744309929.603255,
+      "httpOnly": true,
+      "secure": true,
+      "sameSite": "None"
+    },
+    {
+      "name": "csrf-state-legacy",
+      "value": "\"\"",
+      "domain": "playwright-launchdarkly-login-2a6h5ekup131c0ajka6molsjju.auth.us-east-1.amazoncognito.com",
+      "path": "/",
+      "expires": 1744309929.603325,
+      "httpOnly": true,
+      "secure": true,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "XSRF-TOKEN",
+      "value": "193bcda7-9ede-49c3-99b2-465fbdcc0880",
+      "domain": "playwright-launchdarkly-login-2a6h5ekup131c0ajka6molsjju.auth.us-east-1.amazoncognito.com",
+      "path": "/",
+      "expires": -1,
+      "httpOnly": true,
+      "secure": true,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "cognito",
+      "value": "\"H4sIAAAAAAAAAAHcACP/yPhZhiIbxvKbjLnI4UwVwxll2NckZpKq7NY5DNxteK18bby2+FZIZ22dTHcdyvSdOFjiYHMXWW15NEHKR6RYvRuVQ3ShFXQk9DoTVN7pV2HWlihMhsD+SIi9jlTIihTVKsEu5joZo+5bB3yS4Z/xWKXKnP/T0lx49zL/maK9CmKJ5tMrZnxkb5MEhZ4OX0MeTjhbKsn1wC1nbeCvOuqEakwS7mmE+06vXaU6HCdy9U5tqIf3j9vuqdREKg2buTSLvCbwMH/RgkdIc5c2WbgZILuGhBNSOnWiDMrBXdz7rVbcAAAA.H4sIAAAAAAAAAAEgAN//coiiBZ2XLqfeU0KuJMQKysVTC/+T4QuejM4f9qkkdN+1hQVPIAAAAA==.4\"",
+      "domain": ".playwright-launchdarkly-login-2a6h5ekup131c0ajka6molsjju.auth.us-east-1.amazoncognito.com",
+      "path": "/",
+      "expires": 1744313230.401326,
+      "httpOnly": true,
+      "secure": true,
+      "sameSite": "Lax"
+    }
+  ],
+  "origins": [
+    {
+      "origin": "http://localhost:5000",
+      "localStorage": [
+        {
+          "name": "ld:$anonUserId",
+          "value": "694ce780-1639-11f0-9356-03029ca20c67"
+        },
+        {
+          "name": "ld:6638280397c1bc569aea5f3f:eyJhbm9ueW1vdXMiOnRydWUsImtpbmQiOiJ1c2VyIiwia2V5IjoiNjk0Y2U3ODAtMTYzOS0xMWYwLTkzNTYtMDMwMjljYTIwYzY3In0=",
+          "value": "{\"clear-data-button\":{\"version\":172,\"flagVersion\":3,\"value\":true,\"variation\":0,\"trackEvents\":false},\"login-page\":{\"version\":172,\"flagVersion\":5,\"value\":true,\"variation\":0,\"trackEvents\":false},\"site-under-maintenance-banner\":{\"version\":172,\"flagVersion\":5,\"value\":\"SCHEDULED\",\"variation\":1,\"trackEvents\":false},\"SHOW_SUBMISSION_LOADER\":{\"version\":172,\"flagVersion\":3,\"value\":true,\"variation\":0,\"trackEvents\":false},\"uat-hide-mmdl-banner\":{\"version\":172,\"flagVersion\":81,\"value\":\"on\",\"variation\":1,\"trackEvents\":false},\"COMPLETE_INTAKE\":{\"version\":172,\"flagVersion\":3,\"value\":false,\"variation\":1,\"trackEvents\":false},\"complete-intake\":{\"version\":172,\"flagVersion\":4,\"value\":true,\"variation\":0,\"trackEvents\":false},\"toggleFaq\":{\"version\":172,\"flagVersion\":70,\"value\":true,\"variation\":0,\"trackEvents\":false},\"cms-home-page\":{\"version\":172,\"flagVersion\":7,\"value\":\"off\",\"variation\":0,\"trackEvents\":false},\"perform-intake\":{\"version\":172,\"flagVersion\":4,\"value\":false,\"variation\":1,\"trackEvents\":false},\"$schema\":1}"
+        },
+        {
+          "name": "ld:6638280397c1bc569aea5f3f:$diagnostics",
+          "value": "{\"dataSinceDate\":1744309628415,\"droppedEvents\":0,\"eventsInLastBatch\":0,\"streamInits\":[]}"
+        },
+        {
+          "name": "banner",
+          "value": "null"
+        }
+      ]
+    },
+    {
+      "origin": "https://playwright-launchdarkly-login-2a6h5ekup131c0ajka6molsjju.auth.us-east-1.amazoncognito.com",
+      "localStorage": [
+        {
+          "name": "amznfbgid",
+          "value": "X93-7682116-7498278:1744309629"
+        },
+        {
+          "name": "Amazon.AWS.Cognito.ContextData.LS_UBID",
+          "value": "as0ol567gh6hp2o92d3s:1744309629910"
+        }
+      ]
+    }
+  ]
+}
