+++ test/e2e/forms/details/medicaid-spa-detail.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/test";
 
-test.describe("Medicaid SPA - Sub Doc", () => {
+test.describe.skip("Medicaid SPA - Sub Doc", () => {
   test.beforeEach(async ({ page }) => {
     await page.goto("/");
     await page.getByTestId("Dashboard-d").click();
