+++ test/e2e/non-func/perf/index.spec.ts
@@ -1,6 +1,6 @@
 import { test } from "@playwright/test";
 
-import * as routes from "../../fixtures/routes";
+import * as routes from "@/fixtures/routes";
 
 const STATIC_ROUTES = routes.STATIC;
