--- test/e2e/pages/index.ts
@@ -1,2 +1,0 @@
-export * from "./loginPage";
-export * from "./home.page";
