--- test/e2e/selectors/components/index.ts
@@ -1,3 +1,0 @@
-import { AddIssueFormSelectors } from "./AddIssueForm/index";
-
-export { AddIssueFormSelectors };
