--- test/e2e/selectors/index.ts
@@ -1,2 +1,0 @@
-export * from "./navigation/index";
-export * from "./components/index";
