--- test/e2e/submissions/spa/chip/other.spec.ts
@@ -1,3 +1,0 @@
-import { test } from "@playwright/test";
-
-test.describe.skip("Other CHIP SPA", () => {});
