--- test/e2e/submissions/spa/medicaid/other.spec.ts
@@ -1,3 +1,0 @@
-import { test } from "@playwright/test";
-
-test.describe.skip("Other Medicaid SPA", () => {});
