--- test/e2e/submissions/waiver/1915-b/capitated/renewal-waiver.spec.ts
@@ -1,3 +1,0 @@
-import { test } from "@playwright/test";
-
-test.describe.skip("Capitated Renewal Waiver Details", () => {});
