--- test/e2e/submissions/waiver/1915-c/appendix-k.spec.ts
@@ -1,3 +1,0 @@
-import { test } from "@playwright/test";
-
-test.describe.skip("1915(c) Appendix K Details", () => {});
