--- test/e2e/submissions/waiver/temp-extension/temp-ext.spec.ts
@@ -1,3 +1,0 @@
-import { test } from "@playwright/test";
-
-test.describe.skip("Temp Extension Request Details", () => {});
