--- test/e2e/utils/users.ts
@@ -1,4 +1,0 @@
-export const testUsers = {
-  state: "george@example.com",
-  reviewer: "reviewer@example.com",
-};
