+++ test/fixtures/notifs.ts
@@ -1,0 +1,46 @@
+export const notifs = [
+  {
+    notifId: "9e38a211-5fd5-44ff-96cd-6d0a22a73a94",
+    body: "Medicaid Alternative Benefit Plan, Premium and Cost Sharing, and CHIP Eligibility SPA templates and implementation guides are now available in OneMAC. New submissions for these SPA types are submitted through the OneMAC system effective [add date].",
+    header: "MMDL SPA forms available in OneMAC",
+    pubDate: "2025-01-07T12:20:56.160Z",
+    expDate: "2025-02-07T11:21:50.210Z",
+    buttonLink: "/faq/spa-admendments",
+    buttonText: "Go to FAQs",
+    disabled: true,
+  },
+  {
+    notifId: "9e38a211-5fd5-44ff-96cd-6d0a22a73a95",
+    body: "In addition to a refreshed look, OneMAC now brings breadcrumb navigation, drag-and-drop file upload, improved dashboard filtering, and more.",
+    header: "Enhanced experience",
+    pubDate: "2025-01-07T12:20:56.160Z",
+    expDate: "2025-08-07T11:21:50.210Z",
+    buttonLink: "",
+    buttonText: "More details",
+    disabled: true,
+  },
+  {
+    notifId: "9e38a211-5fd5-44ff-96cd-6d0a22a73a96",
+    header: "Latest Updates",
+    pubDate: "2025-01-07T12:20:56.160Z",
+    expDate: "2025-03-07T11:21:50.210Z",
+    body: "Medicaid Alternative Benefit Plan, Premium and Cost Sharing, and CHIP Eligibility SPA templates and implementation guides are now available in OneMAC. New submissions for these SPA types are submitted through the OneMAC system effective [date].",
+    buttonLink: "",
+  },
+  {
+    notifId: "9e38a211-5fd5-44ff-96cd-6d0a22a73a97",
+    header: "Latest Updates",
+    pubDate: "2025-01-06T10:00:00.000Z",
+    expDate: "2025-01-01T00:00:00.000Z",
+    body: "Medicaid Alternative Benefit Plan, Premium and Cost Sharing, and CHIP Eligibility SPA templates and implementation guides are now available in OneMAC. New submissions for these SPA types are submitted through the OneMAC system effective [date].",
+    buttonLink: "",
+  },
+  {
+    notifId: "9e38a211-5fd5-44ff-96cd-6d0a22a73a98",
+    header: "Latest Updates",
+    pubDate: "2025-01-05T09:00:00.000Z",
+    expDate: "2025-01-30T00:00:00.000Z",
+    body: "Medicaid Alternative Benefit Plan, Premium and Cost Sharing, and CHIP Eligibility SPA templates and implementation guides are now available in OneMAC. New submissions for these SPA types are submitted through the OneMAC system effective [date].",
+    buttonLink: "",
+  },
+];
