+++ test/package.json
@@ -9,13 +9,12 @@
     "e2e": "playwright test --project=state-user-chrome",
     "e2e:ui": "playwright test --ui --project=state-user-chrome"
   },
-  "dependencies": {
-    "@axe-core/playwright": "^4.8.3",
-    "@playwright/test": "^1.45.3"
-  },
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@axe-core/playwright": "^4.8.3",
+    "@playwright/test": "^1.45.3",
+    "msw": "^2.7.3",
     "playwright-msw": "^3.0.1"
   }
 }
