+++ test/pages/index.ts
@@ -1,3 +1,3 @@
-// export * from "./login.page";
-// export * from "./home.page";
-// export * from "./faq.page";
+export * from "./faq.page";
+export * from "./home.page";
+export * from "./login.page";
